Refactor jump to bottom button into its own file
Add new design that exposes a simpler API, which prevents errors
like forgetting to update the visibility state of both the
jump to bottom parent view *and* the unread message badge view.
