+++ src/shared/mod.rs
@@ -7,6 +7,7 @@ pub mod clickable_view;
 pub mod helpers;
 pub mod html_or_plaintext;
 pub mod icon_button;
+pub mod jump_to_bottom_button;
 pub mod search_bar;
 pub mod styles;
 pub mod text_or_image;
@@ -24,5 +25,6 @@ pub fn live_design(cx: &mut Cx) {
     html_or_plaintext::live_design(cx);
     adaptive_view::live_design(cx);
     cached_widget::live_design(cx);
-    typing_animation::live_design(cx)
+    typing_animation::live_design(cx);
+    jump_to_bottom_button::live_design(cx);
 }
